+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## Bugfixes
 - Fixed `can_take_damage` entity condition causing a crash.
-- Fixed typos in warning message for `modify_scale` when used on a non living entity.
+- Fixed priority on `modify_scale` Scale Modifiers being the same as a regular scale modifier.
+- Fixed flickering with `modify_scale` powers.
+- Fixed modify scale not stacking with a `delay` field higher than 0 and conditional. #61